# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "生产级 Gin 框架脚手架 API",
        "title": "My-Gin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/admin/unlock-account": {
            "post": {
                "description": "管理员解锁被锁定的用户账号",
                "tags": [
                    "安全管理"
                ],
                "summary": "解锁用户账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户UID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "description": "统一 API 响应格式",
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                },
                "request_id": {
                    "description": "请求ID",
                    "type": "string",
                    "example": "abc123"
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "integer",
                    "example": 1704067200
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}